user management added

+++ app/helpers/exam_centers_helper.rb
@@ -2,8 +2,8 @@ module ExamCentersHelper
 
   def exam_center_action_bar(center)
     buttons = []
-    buttons << edit_exam_center(center)
-    buttons << delete_btn(center)
+    buttons << edit_exam_center(center) if can? :edit, ExamCenter
+    buttons << delete_btn(center) if can? :destroy, ExamCenter
     buttons << machine_group(center) if can? :create, Machine
     content_tag(:div,raw(buttons.join(" ")))
   end
@@ -29,9 +29,9 @@ module ExamCentersHelper
     btn_group += "<button type='button' class='btn btn-primary dropdown-toggle' data-toggle='dropdown'><span class='caret'></span><span class='sr-only'>Toggle Dropdown</span></button>"
     btn_group += "<ul class='dropdown-menu' role='menu>"
     btn_group += "<li></li>"
-    #btn_group += "<li class='divider'></li>"
-    #btn_group += "<li><a href='#'>Machine Status</a></li>"
-    #btn_group += "<li class='divider'></li>"
+    btn_group += "<li class='divider'></li>"
+    btn_group += "<li><a href='#'>Machine Status</a></li>"
+    btn_group += "<li class='divider'></li>"
     btn_group += "</ul>"
     btn_group += "</div>"
   end
